+++ keyboards/splitkb/aurora/sweep/keymaps/flinguenheld/features/combo.c
@@ -1,6 +1,9 @@
 // Copyright 2023 Florent Linguenheld (@FLinguenheld)
 // SPDX-License-Identifier: GPL-2.0-or-later
 
+#include QMK_KEYBOARD_H
+#include "keycodes.h"
+
 /* How to :
         - Add an entry in the enum
         - Create a new sequence of keys
